# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MustaFlow AI API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: updateProject
      tags: [projects]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Deleted
...
  /projects/{id}/workflows/{name}/run:
    post:
      operationId: runWorkflow
      tags: [workflows]
      summary: Execute a named workflow in the project container
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: name, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Workflow ran (check `ok` for success)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowRunResponse"
        "404":
          description: Workflow not found
        "409":
          description: Project container is not running
...
  /projects/{id}/container/exec:
    post:
      operationId: execInContainer
      tags: [projects]
      summary: Execute a shell command inside the running container
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContainerExecInput"
      responses:
        "200":
          description: Command output
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContainerExecResult"
        "409":
          description: Container not running
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    Project:
      type: object
      required: [id, name, kind, status, agentMode, createdAt, updatedAt]
      properties:
        id: { type: integer }
        workspaceId: { type: ["integer", "null"] }
        name: { type: string }
        description: { type: ["string", "null"] }
        kind:
          type: string
          enum:
            [
              web,
              mobile-ios,
              mobile-android,
              mobile-cross,
              fullstack,
              dashboard,
              api,
              design,
              automation,
              marketplace,
              chatbot,
              spreadsheet,
            ]
        platform:
          type: string
          enum: [web, ios, android, cross]
        status:
          type: string
          enum: [draft, building, testing, published, failed]
        agentMode:
          type: string
          enum: [lite, eco, power, pro]
        policyStrictness:
          type: string
          enum: [safe, standard, permissive]
          description: "Controls the agent loop's command policy. safe = legacy whitelist; standard (default) = broad allow + deny-list; permissive = admin-only, skips registry allowlist."
        deletedAt:
          {
            type: ["string", "null"],
            format: date-time,
            description: "Soft-delete timestamp. Present only on /projects/trash responses; null/omitted for active projects.",
          }
        lastTaskSummary: { type: ["string", "null"] }
        summary: { type: ["string", "null"] }
        publicSlug: { type: ["string", "null"] }
        siteTitle: { type: ["string", "null"] }
        metaDescription: { type: ["string", "null"] }
        themeColor: { type: ["string", "null"] }
        mobilePreviewUrl: { type: ["string", "null"] }
        containerId:
          {
            type: ["string", "null"],
            description: "Fly.io Machine ID for this project's dev container. Null = not provisioned.",
          }
        containerStatus:
          type: string
          enum: [stopped, starting, running, hibernated, error]
          description: "Current dev container lifecycle state."
        containerUrl:
          {
            type: ["string", "null"],
            description: "Proxy URL to reach the container's dev server.",
          }
        prodContainerId:
          {
            type: ["string", "null"],
            description: "Fly.io Machine ID for the live production container. Null = not deployed.",
          }
        prodContainerStatus:
          type: string
          enum: [stopped, starting, running, hibernated, error, deploying]
          description: "Current production container lifecycle state."
        prodContainerUrl:
          {
            type: ["string", "null"],
            description: "Proxy URL to reach the production container. Null = not deployed.",
          }
        healthScore:
          type: ["integer", "null"]
          minimum: 0
          maximum: 100
          description: "0–100 score derived from generated file quality: accessibility (alt/ARIA/semantic), SEO (title/description/h1), performance (script count/lazy), security (no eval/document.write). Null if no files built yet."
        defaultAgent:
          type: string
          enum: [planning, task, main]
          description: "User's preferred agent for this project. planning = Planning Agent, task = Task Agent (staging gate), main = Main Agent (direct fast edit)."
        projectFormat:
          type: string
          enum: [static-html, react-vite]
          description: "Builder output format. static-html = CDN-based single HTML blob (legacy). react-vite = multi-file React + Vite npm project (default for new web projects)."
        stack:
          type: string
          enum: [react-vite, nextjs, node-api, python-flask, python-fastapi]
          description: "Technology stack chosen at project creation. Immutable — duplicate the project to change stack."
        dbProvider:
          type: string
          enum: [none, postgres, sqlite]
          description: "Which database engine is provisioned for this project. none = no DB."
        dbStatus:
          type: string
          enum: [none, provisioning, connected, error]
          description: "Current database lifecycle state."
        dbConnectionId:
          type: ["string", "null"]
          description: "Opaque identifier for the provisioned database (e.g. Neon project ID). Null = not provisioned."
        blockPublishOnCritical:
          type: boolean
          description: "When true, the publish readiness gate blocks publishing if any unresolved critical (error-severity) security findings exist from the latest check run. Default true."
        dismissedFindingHashes:
          type: array
          items: { type: string }
          description: "Array of dismissed finding keys (file:line:message). Dismissed findings are excluded from the publish security gate."
          nullable: true
        autoFixOnCheckFailure:
          type: boolean
          description: "When true, a refine task is automatically enqueued after any build that has failing checks. Auto-fix fires at most once per build and never on auto-fix tasks themselves."
        autoFixWarningsAfterBuild:
          type: boolean
          description: "When true, the build pipeline runs project-wide ESLint auto-fix at the end of every successful build (before the version snapshot is taken). Default false."
        architectReviewEnabled:
          type: boolean
          description: "When true, a second-opinion architect review runs after every successful build/refine. Critical or fail verdicts trigger one auto-fix turn. Charged as a separate flat fee (2 credits). Default true — opt-out per project."
        e2eEnabled:
          type: boolean
          description: "When true, the agentic builder automatically runs Playwright smoke E2E after every successful web build, and the run_e2e tool is available to the model. Default true."
        multiplayerEnabled:
          type: boolean
          description: "When true, the project accepts real-time multiplayer presence/edit broadcasts on WS /api/projects/{id}/multiplayer. Default false — opt-in per project."
        redirectWwwApex:
          type: boolean
          description: "When true, requests to www.<apex> are 301-redirected to the apex domain (or vice versa). Only meaningful when both apex and www are attached as project domains."
        scannerSemgrepEnabled:
          type: boolean
          description: "When true, Semgrep SAST runs on every build (Task #545). Default true."
        scannerHoundDogEnabled:
          type: boolean
          description: "When true, the HoundDog secret/PII scanner runs on every build (Task #545). Default false."
        scannerTrivyEnabled:
          type: boolean
          description: "When true, the Trivy CVE/IaC scanner runs on every build (Task #545). Default false."
        stagingPublishedSnapshotId:
          type: ["integer", "null"]
          description: "Version ID currently live on the staging slot. Null = not staged. Set by POST /publish?env=staging, promoted by POST /promote."
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    ProjectUpdate:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        status:
          type: string
          enum: [draft, building, testing, published, failed]
        agentMode:
          type: string
          enum: [lite, eco, power, pro]
        policyStrictness:
          type: string
          enum: [safe, standard, permissive]
        defaultAgent:
          type: string
          enum: [planning, task, main]
          description: "User's preferred agent for this project."
        siteTitle: { type: string }
        metaDescription: { type: string }
        themeColor: { type: string }
        blockPublishOnCritical:
          type: boolean
          description: "Enable or disable the critical-findings publish gate for this project."
        autoFixOnCheckFailure: { type: boolean }
        autoFixWarningsAfterBuild: { type: boolean }
        architectReviewEnabled:
          type: boolean
          description: "Toggle the architect review subagent for this project."
        e2eEnabled: { type: boolean }
        multiplayerEnabled: { type: boolean }
        redirectWwwApex: { type: boolean }
        scannerSemgrepEnabled:
          type: boolean
          description: "Run Semgrep SAST on every build (Task #545). Default true."
        scannerHoundDogEnabled:
          type: boolean
          description: "Run HoundDog secret/PII scanner on every build (Task #545). Default false."
        scannerTrivyEnabled:
          type: boolean
          description: "Run Trivy CVE/IaC scanner on every build (Task #545). Default false."
...
    WorkflowRunResponse:
      type: object
      required: [name, ok, output, command]
      properties:
        name: { type: string }
        ok: { type: boolean }
        output: { type: string }
        command: { type: string }
        cwd: { type: ["string", "null"] }
...
    ContainerExecInput:
      type: object
      required: [command]
      properties:
        command:
          type: array
          items: { type: string }
          description: "Command and arguments (e.g. ['npm', 'install'])"
        workdir: { type: string, description: "Working directory inside container (default /app)" }
    ContainerExecResult:
      type: object
      required: [ok, output]
      properties:
        ok: { type: boolean }
        output: { type: string }
...